Extract entity recognizer
